refactor(ui): move query result presentation helpers
